requirements.txt
pymongo==3.7.2
pandas==0.23.4
python-dotenv==0.10.3
